fixed handling of available and installed index files (now copied by storymaker)

+++ lib/src/main/java/scal/io/liger/IndexManager.java
@@ -66,8 +66,14 @@ public class IndexManager {
         Log.d("INDEX", "COPYING JSON FILE " + jsonFileName + " FROM ASSETS TO " + jsonFilePath);
 
         File jsonFile = new File(jsonFilePath + jsonFileName);
-        if (jsonFile.exists()) {
+        // do not replace installed index
+        if (jsonFileName.equals(availableIndexName) && jsonFile.exists()) {
+            Log.d("INDEX", "JSON FILE " + jsonFileName + " ALREADY EXISTS IN " + jsonFilePath + ", DELETING");
             jsonFile.delete();
+        }
+        if (jsonFileName.equals(installedIndexName) && jsonFile.exists()) {
+            Log.d("INDEX", "JSON FILE " + jsonFileName + " ALREADY EXISTS IN " + jsonFilePath + ", NOT COPYING");
+            return;
         }
 
         InputStream assetIn = null;
